+++ tests/test_logger.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+"""Tests for `autosbatch` package."""
+
+from autosbatch.logger import logger
+
+
+def test_logger():
+    """Test logger."""
+    assert logger.name == "autosbatch.logger"